requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.1.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
